Handle nulls, add properties, and improve logging
Updated `RetornarEntidadesDeletarEmCascata` to handle `null` values for `relacoesEmCascata`. Added `IsIgnorarZeroIgualNull` and `IsIgnorarValorAntigoNull` properties to `IEstruturaAlteracaoPropriedade` interface. Modified `IsExisteAlteracaoTipoPrimario` to return early if `propriedadeAlterada.AntigoValor` is `null` and `estruturaAlteracaoPropriedade.IsIgnorarValorAntigoNull` is `true`. Removed an unnecessary line in `EnitdadesAlteracaoPropriedade.cs`. Reformatted parameters in `AtualizarValorPadrao.RetornarValorPadraoPropriedade` for readability. Removed `ContadorParametro` field from `SalvarEntidades` class. Updated `EscreverSaida` in `DepuracaoUtil` to conditionally log SQL parameters only if debugging is attached and logging is enabled. Wrapped `EscreverSaidaInterno` in `DepuracaoUtil` with `#if DEBUG` directive. Added a check in `ContextoDadosExtensao.SalvarPropriedades` to throw an exception if the entity lacks an auto-increment primary key.