Refactor IP handling and introduce public IP retrieval
- Changed return value of `RetornarValorPadrao` in `ValorPadraoIP.cs` to `AplicacaoSnebur.Atual.IpPublico`.
- Introduced new namespace `Snebur` in `ConstantesIP.cs` with static class `ConstantesIP` containing IP-related constants.
- Moved `ConstantesIP` class from `IpUtil.cs` to `ConstantesIP.cs`.
- Added private static field `_ipPublico` to `IpUtil` class.
- Added `RetornarIpPublico` method to `IpUtil` class for retrieving and caching public IP.
- Updated `RetornarIPInformacao` method in `IpUtil` class to use `RetornarIpPublico` when IP is null or empty.

Update Ip property to use public IP address

Modified CabecalhoServicoArquivo.cs in the Snebur.ServicoArquivo namespace to change the assignment of the Ip property from AplicacaoSnebur.Atual.IP to AplicacaoSnebur.Atual.IpPublico. This change updates the property to use a public IP address instead of a private or internal one.