Add CacheAttribute class to Snebur.Comunicacao namespace
A new CacheAttribute class has been introduced in the Snebur.Comunicacao namespace. This class is decorated with the [IgnorarAtributoTS] and [AttributeUsage(AttributeTargets.Method)] attributes, allowing it to be applied to methods. It inherits from System.Attribute and includes a property ExpirarCacheEmMinutos with a public getter and setter. A constructor has been added to initialize this property with an optional parameter. Additionally, using directives for Snebur.Dominio.Atributos and System have been added.